fix: apk v3 signing format (posix tar, no EOA blocks)

+++ scripts/index.sh
@@ -17,8 +17,15 @@ do_index() {
 
     TEMP_INDEX="${DIR}/APKINDEX.tmp"
     apk index --allow-untrusted --arch "$ARCH" --output "$TEMP_INDEX" "$DIR"/*.apk
-    mv "$TEMP_INDEX" "${DIR}/APKINDEX.tar.gz"
-    printf 'index: unsigned %s/APKINDEX.tar.gz\n' "$DIR"
+
+    if [ -n "$PRIVKEY" ] && [ -f "$PRIVKEY" ] && [ -s "$PRIVKEY" ]; then
+        "$SCRIPT_DIR/sign.sh" "$PRIVKEY" "$PUBKEY" "$TEMP_INDEX"
+        mv "$TEMP_INDEX" "${DIR}/APKINDEX.tar.gz"
+        printf 'index: signed %s/APKINDEX.tar.gz\n' "$DIR"
+    else
+        mv "$TEMP_INDEX" "${DIR}/APKINDEX.tar.gz"
+        printf 'index: WARNING: unsigned %s/APKINDEX.tar.gz (no private key)\n' "$DIR"
+    fi
 }
 
 if [ -n "$REPO_DIR" ]; then
